requirements.txt
PyQT5
PyQt5-tools
google-cloud-dialogflow-cx 
gtts 
playsound 
pyaudio